🔄 Created initial file structure for independent systems
Nix Flakes requires that the hardware configuration be stored in the repository, which is a huge L. Because of this, I've restructured the repository to have separate flakes for each of my systems. As I break down the configuration files in the future, some configuration files will likely be shared between systems. When this happens, I'll further organize files.